pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "husqbot"
version = "0.1.0"
description = "RAG system for Husqvarna 701 manuals"
requires-python = ">=3.9"
dependencies = [
    "google-cloud-bigquery>=3.11.4",
    "google-cloud-aiplatform>=1.35.0",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.2",
    "numpy>=1.24.3",
    "pandas>=2.0.3",
    "PyPDF2>=3.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/husqbot"] 